Move models to separate class library
