+++ PartsHoleAPI/Controllers/PartsController.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 
 using PartsHoleAPI.Collections;
-using PartsHoleAPI.Models.Interfaces;
+using PartsHoleLib.Interfaces;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
