--- PartsHoleAPI/Models/Interfaces/IBinModel.cs
@@ -1,10 +1,0 @@
-﻿namespace PartsHoleAPI.Models.Interfaces
-{
-    public interface IBinModel : IModel
-    {
-        int Horizontal { get; set; }
-        bool IsBook { get; set; }
-        string Name { get; set; }
-        int Vertical { get; set; }
-    }
-}
