--- PartsHoleAPI/Models/Interfaces/IInvoiceModel.cs
@@ -1,13 +1,0 @@
-﻿using PartsInventory.Models.Enums;
-
-namespace PartsHoleAPI.Models.Interfaces
-{
-    public interface IInvoiceModel : IModel
-    {
-        int OrderNumber { get; set; }
-        List<IPartModel> Parts { get; set; }
-        string? Path { get; set; }
-        decimal SubTotal { get; set; }
-        SupplierType? SupplierType { get; set; }
-    }
-}
