--- PartsHoleAPI/Models/Interfaces/IModel.cs
@@ -1,12 +1,0 @@
-﻿using MongoDB.Bson.Serialization.Attributes;
-using MongoDB.Bson;
-
-namespace PartsHoleAPI.Models.Interfaces
-{
-   public interface IModel
-   {
-      [BsonId]
-      [BsonRepresentation(BsonType.ObjectId)]
-      string Id { get; set; }
-   }
-}
