+++ PartsHoleModelLibrary/Interfaces/IBinModel.cs
@@ -1,4 +1,4 @@
-﻿namespace PartsHoleAPI.Models.Interfaces
+﻿namespace PartsHoleLib.Interfaces
 {
     public interface IBinModel : IModel
     {
