--- PartsHoleModelLibrary/Secrets.cs
@@ -1,8 +1,0 @@
-﻿namespace PartsHoleAPI.Models
-{
-   public class Secrets
-   {
-      public string UserName { get; set; }
-      public string Password { get; set; }
-   }
-}
